change port

+++ api/landscape.py
@@ -1,10 +1,8 @@
 from flask import Blueprint, request, jsonify
 from flask_restful import Api, Resource
-from flask_cors import CORS
 from model.landscape import db, Landscape
 
 landscape_api = Blueprint('landscape_api', __name__, url_prefix='/api')
-CORS(landscape_api) 
 api = Api(landscape_api)
 
 class LandscapeAPI(Resource):
@@ -16,12 +14,15 @@ class LandscapeAPI(Resource):
         data = request.get_json()
         new_landscape = Landscape(data['name'], data['country'], data['city'], data['description'])
         new_landscape.create()
-        return jsonify(new_landscape.read())
+        return jsonify(new_landscape.read()), 201
     
     def delete(self):
         data = request.get_json()
         landscape = Landscape.query.get(data['id'])
-        landscape.delete()
-        return jsonify(f'Successfully deleted {landscape}')
+        if landscape:
+            landscape.delete()
+            return jsonify(landscape.read()), 200
+        else:
+            return jsonify({"error": "Landscape not found"}), 404
 
 api.add_resource(LandscapeAPI, '/landscapes')
